split component CreateSchedulEntries
